Testing with Sound sensor
Created some test sketches of the sound sensor. Right now it doesn't do too great. It works but it seem that its maximum range is 1000hz. Anything above that, it seems to return to zero and start from there. Still need to integrate it with multitasking.

This commit also includes the addition of the fast fourier transform library that converts the detected sound into a frequency that can be detected. . . supposedly.