Debug and class functions for easy access
I moved all the debug print data out of the class into the debugStuff function. This required creating new methods in the class to allow accessing the stored data in there. These include: getMenuName, getMenuBelongsTo, getSelectedItem. And since the cursor doesn't stay at the bottom of the screen anymore (since it moves to draw the selection rectangle), I also created a variable and method to store and access the y position after the last item listed in a given menu.

Right now all the new stuff listed above is hard coded to show for a specific menu only; the start up menu. I need to figure out a way of passing the menu number integer to the menu class rather than the objects array address.