UI Refactor
It was getting way too out of hand dealing with the UI code I came up with so I decided to start from scratch. I've gained a lot of experience over the weeks so I thought I might start from a blank canvas. So far so good. I've been able to get rid of the hard coded menu numbers and order positions. The menu number will be assigned based on where a menu in the array is. And I don't see much use for position anymore since I can just use the location in the array again as the indicator of the submenus listing order.