OpenMenu setting and getting for UI Refactor
Adafruit libraries got updated.

Got rid of the menuNum variable as it wasn't being used and any time I tried to make use of it, it got very complicated very quickly. So moving forward the menuNumber (actually, the index number of the menu objects) will be rightfully referred to as index numbers.

Added getters and setters for the openMenu variable. So the first thing the draw method will do is set the open menu based on the menu name. The function will look up which index number matches with the provided menu name and spit out the index number as though it's a menu number. For all intents and purposes, it is.