Troubleshooting GoDown method
the setter in the GoDown method is somehow making the last child in the starting menu disappear. Only seems to happen when there's more than 3 children on a page and only seems to happen when that setter is called. Also, no idea how that's even possible because that setter isn't even called until AFTER all the children are calculated and listed on the screen. . . maybe it's a memory issue?? It's so odd. I can't seem to track down how that's possible.