bug hunting
I've mostly done away with the Nav class, still some residual methods in the menu class I need to delete now that nav is gone but they're not being used in the code anymore. In regards to the missing menu names issue, I discovered that the length of a menu name affects whether it gets stored in the _menuName variable or not. What's odd about it is that it doesn't seem related to the length of a specific _menuName, but rather the total of all menu names being stored. I haven't proved this yet. When I changed the length of "hello world" to just "hello" it didn't fix the issue (rather it caused another one) and both "hello" and "body profile" were missing from the LCD and serial monitor, and when i put that menu name back and shortened "body profile", the same thing occurred. When I shortened both, the issue went away.

I noticed that in the DrawChildren() method I was retrieving the _menuName and storing it in a local string, a and printing a to the serial monitor and LCD. when I got rid of the local variable and printed _menuName directly, it resolved the issue. I noticed the issue crop up under the GoDown() method serial print so I did the same. I got rid of the local variable that was storing then printing the _menuName and instead printed it directly. A bit more complicated than that but it worked. What I think is happening is the String data type has some limitations I'm unaware of. I may have to find an alternative.

The code as-is doesn't work yet. I'm in the middle of fixing the issue now that I know the cause.