Touch screen calibration
I've added some calculations to determine what the bounds of a menu item should be when touched. For some reason the touch detection doesn't correspond with the xy values of the calculated bounds. When touching the edge of the header bound, i'm getting 34 but the value that's calculated is 40. Need to figure that out next.