TouchScreen Calibration Testing
A lot going on this time. I was able to adapt a calibration sketch found at Bytes N Bits website^1 (referenced below) to work with my breakout. It involved getting rid of their custom library and integrating the touch library I'm using. Next step is to integrate the calibration sketch into my UI so that presses line up with the menu name locations.

Ref 1: https://bytesnbits.co.uk/arduino-touchscreen-calibration-coding/#1606910525368-25b26a28-b2d4