Nav button fixes
Got select to work properly. Still working on the back button. What's become obvious though, is that I need to clean up the code. There's a lot of code I can rewrite to be a method or function. Like converting a menu name to a number (which I created but didn't implement yet) or a child of matches menu name method? Something to work on AFTER I fix the nav issues.
I also put in a blue screen check. If the screen tries to draw a menu that doesn't exist, it will give an error message.