UI is Operational
Removed some unneeded code but mostly addressed all previously found issues with improper menus being selected. A new method was created to reset the selections each time select or back are clicked.

I can add some additional controls so that when the end of a menu branch is reached, either the back or select button are hidden. I may even add a "return to menu" button.

And of course I can optimize to use the quick draw functions so that the whole screen doesn't flash each time the screen is touched.