Added library and started adding sensors
I've started using the temp sensor and integrating it with the menu class. Still a work in progress but it does work so far. What I want to address though is that I have to explicitly spell out if temperature menu then show temp. Need to find a way to just pass the temp sensor to the temperature menu rather than doing a check.