Fighting ChatGPT
Holy hell did GPT take me on a ride. It REALLY wanted me to use the composition method of class creation rather than the inheritance method. But every single question I had made it back track to the point of just using the inheritance method. What a waste of time.

Anywho, this push focused on separating the SensorSuite base class from the TempHumidSensor class. Right now the type of sensor a particular menu will need to display is passed in via the menu obj initialization as a third parameter. So the check that runs after every draw will see if the sensor parameter is null or not, and if not, it will display the contents of the passed in parameter. The beauty here is that you don't have to program in a check inside the menu class to determine what sensor to display based on the name of the menu. The whole damn point of making the menu obj array dynamic was to avoid having to type out EVERY new menu I decided to add. So, the check will first see if the menu being displayed even has a sensor parameter and if it does, then it will see what that parameter is and then display that one.

Still need to decide how to use the sensorSuite class. Maybe I'll store all the methods responsible for interpreting sensor data from multiple sources inside that class. And the subclasses won't really do anything except have getters. Where would I keep the polling method tho? Per subclass or in the base class? Maybe polling can be in the base class and sensor-specific timing functions like in the ultrasonic sensor can be within their own classes?