Refactoring into .h and .cpp files
Moving classes into their own files as the code was getting really unruly. I'm pretty surprised it worked. Now that the code is all split up, it will be easier to identify ways to better encapsulate code.

I suppose next is to start rebuilding the SensorSuite code. Now I can do it in its own .h and .cpp files to make it easier on myself.