Squashing bug 5
https://github.com/hjsemirdzhyan/Tricorder-Project/issues/5

Issue was that the check to confirm a parent menu had children in order to set its first child as "selected", it first checked that _childrenArray wasn't empty. But because that array is only populated AFTER a menu is rendered, it saw it as empty and therefore setting the parent menu as the "selected" menu. Only after redrawing the parent menu and that check running again does the children array actually have content.