Yet another refactor
The spaghetti code is getting out of hand. I'm using my lost-track situation to redo a bunch of the code. I've already taken care of the calibration related code. Next will need to be the menu class. I think it would be best to empty out all methods not related to ACTUAL menu specific functions. The button related methods should be their own class, for example.