Just doing research
Library update and doing some research for the refactor. I'd like to break out some of my static functions out of the menu classes. They don't need to be in there. They're only there because I couldn't figure out how to retrieve some of the values calculated in the menu class from outside of it. The answer there is Get methods.

Also learned about namespaces and how they can be useful in place of classes. Classes instantiate. Namespaces don't. For example the calibrate and isTouched functions don't need to be in a class. Calibrate only runs once and stores its resulting values and isTouched is class independent.

Key takeaway here is separate .h/.cpp files don't have to always be classes. They can be namespaces or even free floating functions.